Remove old Collect interface, replace with new one (API break)
