+++ src/telemetry/CMakeLists.txt
@@ -1,7 +1,6 @@
 zeek_add_subdir_library(
     telemetry
     SOURCES
-    Collect.cc
     Counter.cc
     Gauge.cc
     Histogram.cc
